T09/feature/refactoring

+++ src/components/TodoLayout/TodoList/components/Todo/Todo.tsx
@@ -1,8 +1,9 @@
-import { Todo as TodoProps } from "@/common/types/Todo.ts";
-import { Input } from "@/components/ui/Input/Input.tsx";
-import { useAppDispatch } from "@/hooks/useAppDispatch.ts";
-import { useModalState } from "@/hooks/useModalState.ts";
-import { deleteTodo, setCurrentTodo, updateStatusTodo } from "@/store/actions/todoActionCreators.ts";
+import { NotificationType, TodoConfirmMessages, TodoNotificationMessages } from "@constants";
+import { Input } from "@components";
+import { useAppDispatch, useModalState } from "@hooks";
+import { deleteTodo, setCurrentTodo, setNotification, updateStatusTodo } from "@store";
+import { Todo as TodoProps } from "@types";
+import { checkOnCurrentExpirationDate } from "@utils";
 import cn from "classnames";
 import { FC, useState } from "react";
 import { AiOutlineInfoCircle } from "react-icons/ai";
@@ -13,26 +14,33 @@ import styles from "./Todo.module.scss";
 
 export const Todo: FC<TodoProps> = ({ todoTitle, createdDate, expirationDate, todoId, isCompleted }) => {
   const [isShowInfo, setIsShowInfo] = useState(false);
-  const setModalActive = useModalState("editTodoModal")[1];
+  const setEditModalActive = useModalState("editTodoModal")[1];
+  const setConfirmModalActive = useModalState("confirmModal")[1];
   const dispatch = useAppDispatch();
 
-  const handleChangeStatusTodo = () => {
+  const handleChangeStatusTodo = (): void => {
     dispatch(updateStatusTodo(todoId));
   };
 
-  const handleClickDeleteTodo = () => {
-    dispatch(deleteTodo(todoId));
+  const handleClickDeleteTodo = (): void => {
+    setConfirmModalActive(true, {
+      confirmCallback: () => {
+        dispatch(deleteTodo(todoId));
+        dispatch(setNotification({ title: TodoNotificationMessages.DELETE_TODO, type: NotificationType.SUCCESS }));
+      },
+      message: TodoConfirmMessages.DELETE_TODO,
+    });
   };
 
-  const handleClickEditTodo = () => {
+  const handleClickEditTodo = (): void => {
     if (!isCompleted) {
-      setModalActive(true);
+      setEditModalActive(true);
       dispatch(setCurrentTodo({ todoId, todoTitle, expirationDate, createdDate, isCompleted }));
     }
   };
 
   return (
-    <li className={styles.todoContainer}>
+    <li className={cn(styles.todoContainer, !checkOnCurrentExpirationDate(expirationDate) && styles.expiredTodoContainer)}>
       <div>
         <div className={styles.todoContent}>
           <label className={styles.todoCheck}>
